backend/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
djoser==2.2.0
Pillow==10.0.0
django-filter==23.2
